VP-1099 Moved person permissions to their own file, adding tests for each role
