+++ server/api/person/__tests__/person.spec.js
@@ -2,12 +2,9 @@ import test from 'ava'
 import request from 'supertest'
 import { server, appReady } from '../../../server'
 import Person from '../person'
-import { jwtData, jwtDataDali } from '../../../middleware/session/__tests__/setSession.fixture'
+import { jwtData } from '../../../middleware/session/__tests__/setSession.fixture'
 import MemoryMongo from '../../../util/test-memory-mongo'
 import people from '../__tests__/person.fixture'
-import sinon from 'sinon'
-import { TOPIC_PERSON__CREATE } from '../../../services/pubsub/topic.constants'
-import PubSub from 'pubsub-js'
 import objectid from 'objectid'
 test.before('before connect to database', async (t) => {
   try {
@@ -37,33 +34,6 @@ test('verify fixture database has people', async t => {
   })
 })
 
-test('list everyone - Anon user', async t => {
-  const resAnon = await request(server)
-    .get('/api/people')
-    .set('Accept', 'application/json')
-  t.is(resAnon.status, 403) // Return data response will be undefined for 403 response
-})
-
-test('list everyone - General User', async t => {
-  // general user
-  const res = await request(server)
-    .get('/api/people')
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtDataDali.idToken}`])
-    .expect(200)
-  t.is(res.body.length, t.context.people.length)
-})
-
-test('list everyone - admin user', async t => {
-  // admin user
-  const resAdmin = await request(server)
-    .get('/api/people')
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtData.idToken}`])
-    .expect(200)
-  t.is(resAdmin.body.length, t.context.people.length)
-})
-
 test('list everyone - bad filter', async t => {
   // bad request
   const resBad = await request(server)
@@ -85,26 +55,6 @@ test('list everyone - filtered, sorted, produced', async t => {
   t.is(resFilter.body[0].phone, testers[testers.length - 1].phone)
 })
 
-test('Get person by id', async t => {
-  const p = t.context.people[2] // Testy A.
-  const id = p._id
-  // fetch the person - no auth
-  await request(server)
-    .get(`/api/people/${id}`)
-    .set('Accept', 'application/json')
-    .expect(403)
-
-  // fetch the person - with auth
-  const res = await request(server)
-    .get(`/api/people/${id}`)
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtData.idToken}`])
-    .expect(200)
-
-  t.is(res.body.name, p.name)
-  t.is(res.body.email, p.email)
-})
-
 test('Get person who doesnt exist', async t => {
   // fetch the person - no auth
   const notPersonId = objectid().toString()
@@ -113,102 +63,6 @@ test('Get person who doesnt exist', async t => {
     .set('Accept', 'application/json')
 
   t.is(res.status, 404)
-})
-
-test.serial('create a new person', async t => {
-  t.plan(7)
-
-  // subscribe to published new person messages
-  const spy = sinon.spy()
-  const clock = sinon.useFakeTimers()
-  PubSub.subscribe(TOPIC_PERSON__CREATE, spy)
-  const p = {
-    name: 'Addy McAddFace',
-    // nickname: 'Addy',
-    // phone: '123 456789',
-    email: 'addy@omgtech.co.nz'
-    // role: ['tester'],
-    // tags: ['tag1', 'tag2', 'tag3']
-  }
-
-  try {
-    // anon user can add a new person
-    const res = await request(server)
-      .post('/api/people')
-      .send(p)
-      .set('Accept', 'application/json')
-      .expect(200)
-
-    // confirm message published, sub called
-    t.is(spy.callCount, 0)
-    clock.tick(1)
-    t.is(spy.callCount, 1)
-    clock.restore()
-
-    // can find by id
-    const id = res.body._id
-    const foundId = await Person.findById(id).exec()
-    t.is(foundId.name, p.name)
-
-    // can find by email
-    const foundEmail = await Person.findOne({ email: p.email }).exec()
-    t.is(foundEmail.name, p.name)
-
-    // person has been given the default image
-    t.is(foundEmail.imgUrl, '/static/img/person/person.png')
-
-    // get the new person
-    const resPerson = await request(server)
-      .get(`/api/people/${id}`)
-      .set('Accept', 'application/json')
-      .set('Cookie', [`idToken=${jwtData.idToken}`])
-      .expect(200)
-    t.is(resPerson.body.name, p.name)
-    t.is(resPerson.body.email, p.email)
-
-    // clean up
-    await Person.deleteOne({ email: p.email })
-  } catch (err) {
-    console.error(err)
-  }
-})
-
-test.serial('Update people', async t => {
-  // update the person
-  const p = t.context.people[0]
-  const id = p._id
-  p.phone = '000 0000 000'
-
-  // anon user cannot update
-  await request(server)
-    .put(`/api/people/${id}`)
-    .send(p)
-    .set('Accept', 'application/json')
-    .expect(403)
-
-  // bad id cannot update, not found
-  await request(server)
-    .put('/api/people/notapersonid')
-    .send(p)
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtDataDali.idToken}`])
-    .expect(404)
-
-  // wrong id cannot update, bad request
-  await request(server)
-    .put('/api/people/5d9fe53e4eb179218c8d1d2b')
-    .send(p)
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtDataDali.idToken}`])
-    .expect(404)
-
-  const resUpdated = await request(server)
-    .put(`/api/people/${id}`)
-    .send(p)
-    .set('Accept', 'application/json')
-    .set('Cookie', [`idToken=${jwtData.idToken}`])
-    .expect(200)
-  t.is(resUpdated.body.phone, p.phone)
 })
 
 test.serial('Should correctly add a person and sanitise inputs', async t => {
@@ -240,21 +94,6 @@ test.serial('Should correctly add a person and sanitise inputs', async t => {
   t.deepEqual(p.pronoun, savedPerson.pronoun)
 
   await Person.deleteOne({ email: p.email })
-})
-
-test('Anon user cannot remove a user', async t => {
-  const p = t.context.people[3]
-
-  // try to delete the record anonymously
-  await request(server)
-    .delete(`/api/people/${p._id}`)
-    .set('Accept', 'application/json')
-    .expect(403)
-
-  // check person is still in the database
-  const queriedPerson = await Person.findOne({ email: p.email }).exec()
-  t.is(queriedPerson.name, p.name)
-  t.is(queriedPerson.email, p.email)
 })
 
 test('Should correctly handle missing inputs', async t => {
